Starting to pull this into something vaguely reasonable
