+++ tickets/src/components/interface/EventStyles.js
@@ -72,18 +72,15 @@ export const Cta = styled.a`
 `
 
 export const Title = styled.h1`
-  ${Media.phone`
-    margin-top 20px;
-  `};
-
-  margin-bottom: 20px;
+  font-family: 'IBM Plex Sans', sans-serif;
   font-style: normal;
-  font-weight: 500;
-  font-size: 24px;
-  font-style: light;
-  line-height: 47px;
-  grid-column: 1 3;
-  color: var(--darkgrey);
+  font-weight: bold;
+  font-size: 40px;
+  line-height: normal;
+  margin-bottom: 0px;
+  ${Media.nophone`
+    padding-left: 20px;
+  `}
 `
 
 export const TextArea = styled.textarea`
@@ -106,4 +103,57 @@ export const Input = styled.input`
   padding: 10px;
   font-size: 16px;
   color: var(--darkgrey);
+`
+
+export const Date = styled.h2`
+  font-family: 'IBM Plex Sans', sans-serif;
+  font-style: normal;
+  font-weight: 600;
+  font-size: 24px;
+  color: var(--red);
+  margin-top: 0px;
+  margin-bottom: 0px;
+  ${Media.nophone`
+    padding-left: 20px;
+  `}
+`
+
+export const Time = styled.span`
+  border: 0;
+  border-left: var(--grey) solid 2px;
+  padding-left: 10px;
+  margin-left: 10px;
+`
+
+export const Location = styled.p`
+  font-family: 'IBM Plex Sans', sans-serif;
+  font-size: 16px;
+  margin: 0px;
+  padding-left: 20px;
+`
+
+export const Link = styled.li`
+  margin-top: 15px;
+  font-weight: 200;
+  list-style: none;
+  background: url(${props => props.icon}) no-repeat;
+  padding-left: 40px;
+`
+
+export const Links = styled.ul`
+  font-size: 24px;
+  padding: 0px;
+`
+
+export const DescriptionPara = styled.p`
+  margin-bottom: 1em;
+`
+
+export const DescriptionWrapper = styled.div`
+  font-size: 24px;
+  font-family: 'IBM Plex Sans', sans-serif;
+  p {
+    padding: 0.5em 0;
+    margin: 0px;
+  }
 `
